Parallax Update
Added first background with Parallax. Speed is fully customizable and The Camera is actually moving. Every time it moves, it moves the character with it though, this has to be done to all movin entities or they will not move with the character. Still contemplating if this is the right choice to make, but until we have more this is fine.

DEBUG INPUTS
RIGHT = RUN_RIGHT
DOWN = IDLE

Parallax is checking for WarriorGirls' state. If she is running right and NOT in a fight, then the background will move. Otherwise warriorgirl could not move towards enemies without the background moving.